requirements.txt
requests
nose
python-dotenv
boto3==1.14.50
awscli==1.18.127
flask
jsonify
pytest

# grpc service package
grpcio-tools ~= 1.30